move encryption related stuff to encryptionmanager
